third_party/blink/web_tests/fast/dom/HTMLProgressElement/progress-element-with-child-crash-expected.txt

CONSOLE WARNING: An <input> tag was parsed within an open <select> tag which caused a </select> end tag to automatically be inserted. Please add the missing </select> end tag, since this behavior may change in future versions of this browser, possibly causing breakage on this page.
Test for Bug 48019. It is OK not to crash.
